third_party/WebKit/LayoutTests/http/tests/inspector/cache-storage/cache-data-expected.txt

Tests that cache data is correctly populated in the Inspector.

Dumping CacheStorage tree:
    (empty)
Dumping CacheStorage tree:
    cache: testCache1 - http://127.0.0.1:8000
        (cache empty)
    cache: testCache2 - http://127.0.0.1:8000
        (cache empty)
Dumping CacheStorage tree:
    cache: testCache1 - http://127.0.0.1:8000
        'http://fake.request.com/1': 'OK'
        'http://fake.request.com/2': 'Not Found'
    cache: testCache2 - http://127.0.0.1:8000
        'http://fake.request2.com/1': 'OK'
        'http://fake.request2.com/2': 'Not Found'
